TimerProcess: make some variables uint8_t
save a couple of bytes

+++ libraries/AP_PeriodicProcess/AP_PeriodicProcessStub.h
@@ -9,13 +9,13 @@
 class AP_PeriodicProcessStub : public AP_PeriodicProcess
 {
     public:
-        AP_PeriodicProcessStub(int period = 0);
+        AP_PeriodicProcessStub(uint8_t period = 0);
         void init( Arduino_Mega_ISR_Registry * isr_reg );
         void register_process(ap_procedure proc);
         void set_failsafe(ap_procedure proc);
         static void run(void);
     protected:
-        static int  _period;
+        static uint8_t _period;
         static void (*_proc)(void);
         static void (*_failsafe)(void);
 };
